improved test running scripts

+++ tests/run_test.sh
@@ -11,16 +11,24 @@ OUTPUT_FILE=".test_result.shazam.tmp"
 TEST_LOG_FILE="test.last.log.txt"
 
 echo -n "[1] Compiling..."
+
 $COMPILER -std=$CPP_VERSION $TESTFILE -o $OUTPUT_FILE
-echo " Done!"
-echo ""
-echo -n "[2] Running tests..."
-./$OUTPUT_FILE $ARGS > $TEST_LOG_FILE
-echo " Done!"
-echo ""
-echo -n "[3] Cleaning..."
-rm $OUTPUT_FILE
-echo " Done!"
-echo ""
-echo -n "[4] Result: "
-cat $TEST_LOG_FILE
+
+if [ -f $OUTPUT_FILE ]
+then
+    echo " Done!"
+    echo ""
+    echo -n "[2] Running tests..."
+    ./$OUTPUT_FILE $ARGS > $TEST_LOG_FILE
+    echo " Done!"
+    echo ""
+    echo -n "[3] Cleaning..."
+    rm $OUTPUT_FILE
+    echo " Done!"
+    echo ""
+    echo -n "[4] Result: "
+    cat $TEST_LOG_FILE
+else
+    echo ""
+    echo "Error!!"
+fi
